refactor: Update DataTable components to improve code cleanliness and remove unused components
The code changes refactor the DataTable components to improve code cleanliness and remove unused components. This includes updating the UserAvatar component to use the DropdownMenu instead of the deprecated Popover, updating the Sidebar component to use the IFCChildren interface instead of the deprecated ISidebar interface, updating the MainTable component to use the Expense type instead of the deprecated Payment type and adding a generateId function, and updating the HeadSection component to use the IFCChildren interface instead of the deprecated IHeadSectionChildren interface. Additionally, the unused Avatar and FunctionComponents interfaces are removed. These changes improve code consistency, reduce unnecessary complexity, and ensure the correct UI elements and types are used throughout the components.