feat: Add ComboBox component for dropdown selection
This commit adds the ComboBox component to the project. The ComboBox component is a reusable UI component that provides a dropdown selection functionality. It allows users to select an option from a list of predefined options. This addition enhances the user interface by providing an intuitive and interactive way for users to make selections.