Refactor course creation to check for existing course before creating
This commit refactors the course creation process in the CoursesService class. It adds a check to see if a course with the same course number already exists before creating a new course. If a course already exists, the creation process is aborted and an appropriate error message is returned. This change helps to prevent duplicate courses from being created.